+++ Bot/src/logger/logger.py
@@ -1,0 +1,3 @@
+from BetterPythonLogger import Logger
+
+logger = Logger.Logger("Bot")