+++ Bot/src/parserTwitch.py
@@ -1,4 +1,4 @@
-
+from src.logger.logger import logger
 
 class ParserTwitch():
     def __init__(self):
@@ -15,7 +15,12 @@ class ParserTwitch():
         return dictionnary
     
     def parse(self, message):
-        print(f"Message avant de split {message}")
+        # print(f"Message avant de split {message}")
+        if message.startswith("PING"):
+            logger.warning("PING received from Twitch, awaiting PONG...")
+            result = {"command": "PING"}
+            return result
+        
         messages = message.strip().split(":")
         # print("Messages splités: ")
         result = {}