+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# [1.3.0] - 21/03/2024
+## BOT:
+    - Use BetterPythonLogger library
+
+# [1.2.0] - 20/03/2024
+## BOT:
+    - Use logger instead of print
+
+# [1.1.0] - 19/03/2024
+## BOT:
+    - Auto connection token by launching an http server
+
 # [1.0.0] - 18/03/2024
 # - Initial release
 ## BOT: